requirements.txt

rdkit-pypi==2024.03.5
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.4
pyarrow==16.1.0
